Add initial optimization changes
