+++ package.json
@@ -27,6 +27,7 @@
 		"react-modal": "^3.13.1",
 		"react-reveal": "^1.2.2",
 		"react-router-dom": "^5.2.0",
+		"react-spinners": "^0.11.0",
 		"rxjs": "^7.1.0",
 		"styled-components": "^5.3.0",
 		"web3": "^1.3.6",
