+++ src/components/contracts/raptor.ts
@@ -1,5 +1,5 @@
-import {Wallet} from "../wallet";
-import {Contract} from "web3-eth-contract";
+import {Wallet} from '../wallet';
+import {Contract} from 'web3-eth-contract';
 // import { ethers } from 'ethers';
 import * as web3 from 'web3-utils';
