+++ src/components/pages/aboutComponent.css
@@ -23,9 +23,15 @@
 	color: var(--primary);
 }
 
+.about-container #sidebar h1 {
+	color: var(--dark);
+	font-size: x-large;
+	padding-top: 0px;
+}
+
 .about-container #team-members .photo {
-	width: 240px;
-	height: 260px;
+	width: 250px;
+	height: 250px;
 	margin: 30px 0;
 	max-width: 250px;
 	align-self: center;
