+++ src/components/pages/aboutComponent.tsx
@@ -1,9 +1,11 @@
-import * as React from "react";
-import { withTranslation, WithTranslation, TFunction } from "react-i18next";
-import { BaseComponent } from "../shellInterfaces";
+import * as React from 'react';
+
+import { withTranslation, WithTranslation, TFunction } from 'react-i18next';
+import { BaseComponent } from '../shellInterfaces';
 import { Fade, Slide } from 'react-reveal';
-import { pulse } from "react-animations";
-import styled, { keyframes } from "styled-components";
+import { pulse } from 'react-animations';
+import styled, { keyframes } from 'styled-components';
+
 import './aboutComponent.css';
 
 export type AboutProps = {};
@@ -84,7 +86,7 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 					{ text: t('about.our_roadmap.q1_2022.line2'), done: "wip", },
 				],
 			},
-		]
+		];
 
 		return <div className="about-container">
 			<div className="d-flex flex-row container">
@@ -141,9 +143,9 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 								</div>
 								<div className="row">
 									<div className="col-md-6 d-flex flex-column item">
-										<img className="photo" src="images/team/ismail.png" />
+										<img className="photo" src="images/team/ismail.png" alt="raptor-team-founder" />
 										<a href="https://www.linkedin.com/in/ismail-van-essen-b6936628/" target="_blank">
-											<h3 className="name">{t('about.team.founder.name')}</h3>
+											<h2 className="name">{t('about.team.founder.name')}</h2>
 										</a>
 										<p className="title">{t('about.team.founder.title')}</p>
 										<p className="description">{t('about.team.founder.desc1')}</p>
@@ -153,9 +155,9 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 										<p className="description">{t('about.team.founder.desc5')}<br /><br /></p>
 									</div>
 									<div className="col-md-6 d-flex flex-column item">
-										<img className="photo" src="images/team/jonathan.png" />
+										<img className="photo" src="images/team/jonathan.png" alt="raptor-team-vp-marketing" />
 										<a href="http://linkedin.com/in/jonathan-g-51bb35109" target="_blank">
-											<h3 className="name">{t('about.team.vp_marketing.name')}</h3>
+											<h2 className="name">{t('about.team.vp_marketing.name')}</h2>
 										</a>
 										<p className="title">{t('about.team.vp_marketing.title')}</p>
 										<p className="description">{t('about.team.vp_marketing.desc1')}</p>
@@ -169,9 +171,9 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 							<Fade bottom>
 								<div className="row">
 									<div className="col-md-12 d-flex flex-column item">
-										<img className="photo" src="images/team/ilija.png" />
+										<img className="photo" src="images/team/ilija.png" alt="raptor-team-vp-security" />
 										<a href="https://www.linkedin.com/in/ilija-icevic-246136195/" target="_blank">
-											<h3 className="name">{t('about.team.vp_security.name')}</h3>
+											<h2 className="name">{t('about.team.vp_security.name')}</h2>
 										</a>
 										<p className="title">{t('about.team.vp_security.title')}</p>
 									</div>
@@ -179,23 +181,23 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 							</Fade>
 							<Fade left>
 								<div className="row">
-									<div className="col-md-4 d-flex flex-column item"><img className="photo" src="images/team/rinor.png" />
+									<div className="col-md-4 d-flex flex-column item"><img className="photo" src="images/team/rinor.png" alt="raptor-team-social-media-manager" />
 										<a href="https://www.linkedin.com/in/rinor-sherifi-b568bb20b/" target="_blank">
-											<h3 className="name">{t('about.team.social_media_manager.name')}</h3>
+											<h2 className="name">{t('about.team.social_media_manager.name')}</h2>
 										</a>
 										<p className="title">{t('about.team.social_media_manager.title')}</p>
 									</div>
 									<div className="col-md-4 d-flex flex-column item">
-										<img className="photo" src="images/team/kyle.png" />
+										<img className="photo" src="images/team/kyle.png" alt="raptor-team-content-manager" />
 										<a href="https://www.linkedin.com/in/kyle-marcos-a52781211/" target="_blank">
-											<h3 className="name">{t('about.team.content_manager.name')}</h3>
+											<h2 className="name">{t('about.team.content_manager.name')}</h2>
 										</a>
 										<p className="title">{t('about.team.content_manager.title')}</p>
 									</div>
 									<div className="col-md-4 d-flex flex-column item">
-										<img className="photo" src="images/team/romain.png" />
+										<img className="photo" src="images/team/romain.png" alt="raptor-team-graphics-designer" />
 										<a href="https://www.linkedin.com/in/romain-lambert-7a6a657a" target="_blank">
-											<h3 className="name">{t('about.team.graphics_designer.name')}</h3>
+											<h2 className="name">{t('about.team.graphics_designer.name')}</h2>
 										</a>
 										<p className="title">{t('about.team.graphics_designer.title')}</p>
 									</div>
@@ -203,12 +205,12 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 							</Fade>
 							<Fade right>
 								<div className="row">
-									<div className="col-md-6 d-flex flex-column item"><img className="photo blank" src="images/team/blank.png" />
-										<h3 className="name">{t('about.team.software_developer1.name')}</h3>
+									<div className="col-md-6 d-flex flex-column item"><img className="photo blank" src="images/team/blank.png" alt="raptor-team-blank" />
+										<h2 className="name">{t('about.team.software_developer1.name')}</h2>
 										<p className="title">{t('about.team.software_developer1.title')}</p>
 									</div>
-									<div className="col-md-6 d-flex flex-column item"><img className="photo blank" src="images/team/blank.png" />
-										<h3 className="name">{t('about.team.software_developer2.name')}</h3>
+									<div className="col-md-6 d-flex flex-column item"><img className="photo blank" src="images/team/blank.png" alt="raptor-team-blank" />
+										<h2 className="name">{t('about.team.software_developer2.name')}</h2>
 										<p className="title">{t('about.team.software_developer2.title')}</p>
 									</div>
 								</div>
@@ -219,8 +221,8 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 				<Slide right>
 					<div className="d-none d-lg-flex sticky-top col-3" id="sidebar">
 						<PulseDiv className="align-self-start gradient-card primary">
-							<div className="background"><img src="images/logo-rp-u.svg" /></div>
-							<h4><strong>{t('about.card_right.title')}</strong></h4>
+							<div className="background"><img src="images/logo-rp-u.svg" alt="raptor-logo-svg" /></div>
+							<h1><strong>{t('about.card_right.title')}</strong></h1>
 							<p>{t('about.card_right.desc')}</p>
 							<a className="link-dark" href="#mission">{t('about.card_right.mission')}</a>
 							<a className="link-dark" href="#team">{t('about.card_right.team')}</a>
@@ -232,4 +234,4 @@ class aboutComponent extends BaseComponent<AboutProps & WithTranslation, AboutSt
 	}
 }
 
-export default withTranslation()(aboutComponent)
+export default withTranslation()(aboutComponent);
