+++ src/components/pages/homeComponent.css
@@ -7,20 +7,6 @@
 	margin-bottom: 15px;
 }
 
-.home-container #raptor-forest {
-	margin: 10px;
-}
-
-.home-container #raptor-forest p {
-	flex: 0 0 50%;
-}
-
-.home-container #raptor-forest h2 {
-	flex: 1 1 100%;
-	min-width: 100%;
-	margin-bottom: 20px;
-}
-
 .home-container .hero-buttons {
 	margin: 60px 0;
 }
@@ -54,63 +40,6 @@
 	font-size: medium;
 }
 
-.home-container #roadmap h1 {
-	font-size: 56px;
-	text-align: center;
-	margin-bottom: 40px;
-}
-
-.home-container #roadmap h3 {
-	font-size: 40px;
-	font-weight: bold;
-	color: var(--primary);
-}
-
-.home-container #roadmap .row {
-	padding: 0px 20% 0px 20%;
-}
-
-.home-container #roadmap p {
-	font-size: large;
-	text-align: center;
-	margin-bottom: 40px;
-}
-
-.roadmap-list {
-	color: #fff;
-}
-
-.roadmap-list li {
-	margin-bottom: 5px;
-}
-
-.roadmap-list .complete {
-	color: var(--primary-alt);
-}
-
-.roadmap-list .complete::marker {
-	content: ' ✓  ';
-}
-
-.roadmap-content {
-	padding-left: 80px;
-	margin-top: 10px;
-}
-
-.home-container #roadmap h3 {
-	font-weight: bold;
-	font-size: 40px;
-	margin-bottom: 10px;
-}
-
-.line {
-	border-left: 1px solid var(--primary-alt);
-	height: 15px;
-	margin-top: 90px;
-	margin-left: 24px;
-	width: 0px;
-}
-
 @media (max-width: 990px) {
 	.hero-buttons {
 		justify-content: center;
@@ -122,33 +51,5 @@
 
 	.btn-about {
 		margin-left: 10px !important;
-	}
-
-	.btn-roadmap {
-		margin-left: 42%;
-	}
-
-	.roadmap-content {
-		padding-left: 15px;
-	}
-
-	.line {
-		margin-left: 51%;
-	}
-
-	.home-container #roadmap h1 {
-		font-size: 48px;
-		text-align: center;
-		margin-bottom: 20px;
-	}
-
-	.home-container #roadmap .row {
-		padding: 0px 2% 0px 2%;
-	}
-
-	.home-container #roadmap p {
-		font-size: large;
-		text-align: center;
-		margin-bottom: 20px;
 	}
 }
