+++ src/components/pages/lotteryComponent.css
@@ -20,9 +20,15 @@
 	padding: 20px;
 }
 
-.lottery-container .lottery-body h3 {
-	margin-bottom: 20px;
+.lottery-container .lottery-body h1 {
+	font-size: 1.8rem;
 	font-weight: bold;
+	margin-bottom: 20px;
+}
+
+.lottery-container .lottery-body h2 {
+	font-size: 1.2rem;
+	margin-bottom: 8px;
 }
 
 .lottery-info {
