+++ src/components/shellNav.tsx
@@ -1,20 +1,20 @@
 import * as React from 'react';
-import { NavLink } from "react-router-dom";
-import { BaseComponent, IShellPage } from "./shellInterfaces";
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-import './shellNav.css';
-import './shellNav.icons.css';
-import { faBars } from "@fortawesome/free-solid-svg-icons";
-import { faMedium, faTelegram, faTwitter } from "@fortawesome/free-brands-svg-icons";
-import { faFacebook } from "@fortawesome/free-brands-svg-icons/faFacebook";
-import { faInstagram } from "@fortawesome/free-brands-svg-icons/faInstagram";
-import { faReddit } from "@fortawesome/free-brands-svg-icons/faReddit";
-import { faTiktok } from "@fortawesome/free-brands-svg-icons/faTiktok";
-import { faYoutube } from "@fortawesome/free-brands-svg-icons/faYoutube";
-import { faDiscord } from "@fortawesome/free-brands-svg-icons/faDiscord";
+import { NavLink } from 'react-router-dom';
+import { BaseComponent, IShellPage } from './shellInterfaces';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faBars } from '@fortawesome/free-solid-svg-icons';
+import { faMedium, faTelegram, faTwitter } from '@fortawesome/free-brands-svg-icons';
+import { faFacebook } from '@fortawesome/free-brands-svg-icons/faFacebook';
+import { faInstagram } from '@fortawesome/free-brands-svg-icons/faInstagram';
+import { faReddit } from '@fortawesome/free-brands-svg-icons/faReddit';
+import { faTiktok } from '@fortawesome/free-brands-svg-icons/faTiktok';
+import { faYoutube } from '@fortawesome/free-brands-svg-icons/faYoutube';
+import { faDiscord } from '@fortawesome/free-brands-svg-icons/faDiscord';
 import { TFunction, withTranslation, WithTranslation } from 'react-i18next';
 import { supportedLanguages, languageCodeOnly } from '../i18n';
+import './shellNav.css';
+import './shellNav.icons.css';
 
 export type ShellNavProps = {
 	pages: IShellPage[];
@@ -74,31 +74,31 @@ class ShellNav extends BaseComponent<ShellNavProps & WithTranslation, ShellNavSt
 					</ul>
 					<div className="navigation-footer">
 						<div className="social-medias text-center text-lg-left">
-							<a href="https://twitter.com/raptor_token" className="btn-social" target="_blank">
+							<a href="https://twitter.com/raptor_token" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faTwitter} />
 							</a>
-							<a href="https://t.me/PhilosoRaptorToken" className="btn-social" target="_blank">
+							<a href="https://t.me/PhilosoRaptorToken" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faTelegram} />
 							</a>
-							<a href="https://www.facebook.com/raptr.finance" className="btn-social" target="_blank">
+							<a href="https://www.facebook.com/raptr.finance" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faFacebook} />
 							</a>
-							<a href="https://www.instagram.com/raptor_token/" className="btn-social" target="_blank">
+							<a href="https://www.instagram.com/raptor_token/" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faInstagram} />
 							</a>
-							<a href="https://www.reddit.com/r/RaptorToken/" className="btn-social" target="_blank">
+							<a href="https://www.reddit.com/r/RaptorToken/" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faReddit} />
 							</a>
-							<a href="https://www.tiktok.com/@raptor.finance" className="btn-social" target="_blank">
+							<a href="https://www.tiktok.com/@raptor.finance" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faTiktok} />
 							</a>
-							<a href="https://www.youtube.com/channel/UCQ-yByM7ECDvB1CIr_5Pafg" className="btn-social" target="_blank">
+							<a href="https://www.youtube.com/channel/UCQ-yByM7ECDvB1CIr_5Pafg" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faYoutube} />
 							</a>
-							<a href="https://discord.gg/EJwR3pjd9A" className="btn-social" target="_blank">
+							<a href="https://discord.gg/EJwR3pjd9A" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faDiscord} />
 							</a>
-							<a href="https://raptor-finance.medium.com/" className="btn-social" target="_blank">
+							<a href="https://raptor-finance.medium.com/" rel="noreferrer" className="btn-social" target="_blank">
 								<FontAwesomeIcon icon={faMedium} />
 							</a>
 						</div>
@@ -130,4 +130,4 @@ class ShellNav extends BaseComponent<ShellNavProps & WithTranslation, ShellNavSt
 	}
 }
 
-export default withTranslation()(ShellNav)
+export default withTranslation()(ShellNav);
